backend_prueba_tecnica/requirements.txt
djangorestframework-simplejwt
djangorestframework
dj-database-url==2.3.0
Django==5.1.6
psycopg2-binary==2.9.10
google-auth
google-auth-oauthlib
python-decouple==3.8